🎨 style(app): add new file app.context.ts and update css styles This commit adds a new file app.context.ts and updates css styles in several files. The changes include: - Added a new file app.context.ts that exports a React context for the baseUrl. - Updated the Games.module.scss file to add a new css class .hiddenLeaderboardRow. - Updated the LiveRun.module.scss file to add new css classes .splitRowActive and .finalSplitRow. - Updated the User.module.scss file to add a new css class .sameGameRunHeader. - Updated the game-leaderboards.tsx file to use the new css class .hiddenLeaderboardRow. - Updated the layout.tsx file to use the new baseUrl context and to handle isLoading state when fetching patreon data. - Updated the links.tsx file to use the new baseUrl context and to handle isLoading state when fetching patreon data. - Updated the live-user-run.tsx file to
🚀 chore: update react-bootstrap to version 2.7.4
React-bootstrap has been updated to version 2.7.4 to ensure that the application is using the latest version of the library.

🐛 fix(patreon-name.tsx): add isLoading check to usePatreons hook
The usePatreons hook now returns an object with a data and isLoading property. The isLoading property is used to ensure that the patreons data is loaded before rendering the component.

🐛 fix(use-patreons.ts): return useSWR hook directly
The usePatreons hook now returns the useSWR hook directly instead of destructuring the data property.

🐛 fix(compare-splits.tsx): add baseUrl to useContext call
The baseUrl is now retrieved from the AppContext in CompareSplits component.

🐛 fix(user-overview.tsx): add className to thead element
The thead element now has a

🐛 fix(_app.tsx): fix baseUrl variable assignment in getInitialProps
✨ feat(_app.tsx): add AppContext to provide baseUrl to child components
The baseUrl variable was not being assigned correctly in the getInitialProps function of the _app.tsx file. This has been fixed to correctly assign the protocol and host to the baseUrl variable. Additionally, an AppContext has been added to provide the baseUrl to child components.

🐛 fix(change-appearance.tsx): remove unused baseUrl import
The baseUrl variable was being imported from the _app.tsx file but was not being used. This import has been removed.

🐛 fix(index.tsx, live.tsx, patron.tsx, [tournament].tsx): add isLoading check to SWR hooks
The SWR hooks in the index.tsx, live.tsx, patron.tsx, and [tournament].tsx files were not checking for the isLoading state. This has been fixed