🔒 chore(.env): rename TWITCH_OAUTH_CLIENT_ID to NEXT_PUBLIC_TWITCH_OAUTH_CLIENT_ID The environment variable name is updated to be more descriptive and consistent with the naming conventions.
🔥 refactor(topbar.tsx): replace TwitchLoginServer with TwitchLoginButton
The TwitchLoginServer component is removed and replaced with the TwitchLoginButton component. This change improves the consistency of the naming conventions and simplifies the code.

🔒 chore(twitch-oauth.ts, is-channel-live.ts, login-with-twitch.ts): rename TWITCH_OAUTH_CLIENT_ID to NEXT_PUBLIC_TWITCH_OAUTH_CLIENT_ID
The environment variable name is updated to be more descriptive and consistent with the naming conventions.

🔒 chore(layout.tsx, change-appearance.tsx, how-it-works.tsx): add suppressHydrationWarning to html tag
The suppressHydrationWarning attribute is added to the html tag to suppress the warning message that appears when the