Next App Router (#53)
* Start migrating to app router

* Get auth and baseurl and rest working with app router

* Fixed path mappings that broke build

* ignore vendor in eslint

* 🎨 style(layout.tsx): add Material Symbols Outlined font
A new font, Material Symbols Outlined, has been added to the project. This font is used to display icons in the application. The font is loaded via a font-face declaration in the material-symbols-outline.css file. The font is then used in the application by adding the .material-symbols-outlined class to an element.

* [Update] Prepare dark mode for next13 update

* [Bugfix] Load bootstrap before overrides

* Load bootstrap via cdn

* check for undefined string in session

* [Change] Add next-themes for fixing FOUC on initial dark theme load

* fix linter

* made twitch client id public

* 🔒 chore(.env): rename TWITCH_OAUTH_CLIENT_ID to NEXT_PUBLIC_TWITCH_OAUTH_CLIENT_ID
The environment variable name is updated to be more descriptive and consistent with the naming conventions.

🔥 refactor(topbar.tsx): replace TwitchLoginServer with TwitchLoginButton
The TwitchLoginServer component is removed and replaced with the TwitchLoginButton component. This change improves the consistency of the naming conventions and simplifies the code.

🔒 chore(twitch-oauth.ts, is-channel-live.ts, login-with-twitch.ts): rename TWITCH_OAUTH_CLIENT_ID to NEXT_PUBLIC_TWITCH_OAUTH_CLIENT_ID
The environment variable name is updated to be more descriptive and consistent with the naming conventions.

🔒 chore(layout.tsx, change-appearance.tsx, how-it-works.tsx): add suppressHydrationWarning to html tag
The suppressHydrationWarning attribute is added to the html tag to suppress the warning message that appears when the

* Set icons dynamically

* Fixed husky pre-commit hook, moved HelmetProvider to providers.tsx

---------

Co-authored-by: Joey <info@therun.gg>
Co-authored-by: Sebastian Zoglowek <55794780+zoglo@users.noreply.github.com>