Migrate /upload and /upload-key to app router (#59)
* ✨ feat(upload-key): add new route and components for upload key management
The new route and components allow users to manage their upload keys, which are used to automatically upload splits from LiveSplit. Users can now view, copy, and use their upload keys, as well as access instructions on how to set up the LiveSplit component and troubleshoot any issues. This feature enhances the user experience by making it easier to manage and use upload keys.

✨ feat(upload/page.tsx): add new Upload page with drag and drop functionality
🔧 refactor(session.action.ts): update getSession return type to Promise<User> for better type checking
🔥 remove(api/users/[user]/upload-key.ts): delete unused upload-key API endpoint
The new Upload page allows users to upload their splits using a drag and drop interface. The getSession function in session.action.ts has been updated to return a Promise<User> for better type checking and improved code quality. The unused upload-key API endpoint has been removed to keep the codebase clean and maintainable.

* ✨ feat(app/error.tsx): add Error component to handle and display errors
The Error component is added to handle and display errors in the application. It logs the error to the console and provides a "Try again" button to attempt to recover by re-rendering the segment. This improves the user experience by providing a clear indication of errors and a way to recover from them.

---------

Co-authored-by: therun.gg <info@therun.gg>